+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - updated Claude model to 3.7 in default preferred list of models
 - system prompt for MCP agent can be now configured in settings
 - fixed prompt field focus issue after model selection
+- properly showing preferred models in model selector when searching
 
 ## [0.3.1]